#323 Add occurrence ratio to BasicStatistics and CUCCs.

+++ backend/src/main/java/de/metanome/backend/result_postprocessing/result_ranking/UniqueColumnCombinationRanking.java
@@ -92,7 +92,7 @@ public class UniqueColumnCombinationRanking extends Ranking {
     for (ColumnIdentifier column : result.getColumnCombination().getColumnIdentifiers()) {
       occurrences += this.occurrenceMap.get(result.getTableName()).get(column.getColumnIdentifier());
     }
-    result.setOccurrenceRatio(result.getColumnCombination().getColumnIdentifiers().size() / occurrences);
+    result.setOccurrenceRatio((float) result.getColumnCombination().getColumnIdentifiers().size() / occurrences);
   }
 
 }
